Rade/confirm modal

+++ CODE/ZdravoCorpAppTim22/View/Manager/ViewModels/MedicineViewModels/MedicineViewModel.cs
@@ -5,6 +5,7 @@ using ZdravoCorpAppTim22.Controller;
 using ZdravoCorpAppTim22.Model;
 using ZdravoCorpAppTim22.View.Manager.Commands;
 using ZdravoCorpAppTim22.View.Manager.Pages.MedicinePages;
+using ZdravoCorpAppTim22.View.Manager.Views;
 
 namespace ZdravoCorpAppTim22.View.Manager.ViewModels.MedicineViewModels
 {
@@ -39,39 +40,42 @@ namespace ZdravoCorpAppTim22.View.Manager.ViewModels.MedicineViewModels
 
         public void DeleteMedicine(object obj)
         {
-            Medicine medicine = (Medicine)obj;
-            MedicineCollection.Remove(medicine);
-            List<Medicine> medicineToRemove = new List<Medicine>();
-            List<Ingredient> ingredientToRemove = new List<Ingredient>();
-
-            foreach (Ingredient ingredient in medicine.MedicineData.Ingredient)
+            if (ConfirmModal.Show("Are you sure?"))
             {
-                ingredientToRemove.Add(ingredient);
-            }
+                Medicine medicine = (Medicine)obj;
+                MedicineCollection.Remove(medicine);
+                List<Medicine> medicineToRemove = new List<Medicine>();
+                List<Ingredient> ingredientToRemove = new List<Ingredient>();
 
-            foreach (Medicine m in MedicineController.Instance.GetAll())
-            {
-                if(m.MedicineData.Id == medicine.MedicineData.Id)
+                foreach (Ingredient ingredient in medicine.MedicineData.Ingredient)
                 {
-                    medicineToRemove.Add(m);
-                    
+                    ingredientToRemove.Add(ingredient);
                 }
+
+                foreach (Medicine m in MedicineController.Instance.GetAll())
+                {
+                    if(m.MedicineData.Id == medicine.MedicineData.Id)
+                    {
+                        medicineToRemove.Add(m);
+                    
+                    }
+                }
+                foreach(Ingredient ingredient in ingredientToRemove)
+                {
+                    ingredient.MedicineData = null;
+                    IngredientController.Instance.DeleteByID(ingredient.Id);
+                }
+                foreach(Medicine m in medicineToRemove)
+                {
+                    MedicineController.Instance.DeleteByID(m.Id);
+                }
+                medicine.MedicineData.RemoveAllIngredient();
+                if(medicine.MedicineData.Approval != null)
+                {
+                    ApprovalController.Instance.DeleteByID(medicine.MedicineData.Approval.Id);
+                }
+                MedicineDataController.Instance.DeleteByID(medicine.MedicineData.Id);
             }
-            foreach(Ingredient ingredient in ingredientToRemove)
-            {
-                ingredient.MedicineData = null;
-                IngredientController.Instance.DeleteByID(ingredient.Id);
-            }
-            foreach(Medicine m in medicineToRemove)
-            {
-                MedicineController.Instance.DeleteByID(m.Id);
-            }
-            medicine.MedicineData.RemoveAllIngredient();
-            if(medicine.MedicineData.Approval != null)
-            {
-                ApprovalController.Instance.DeleteByID(medicine.MedicineData.Approval.Id);
-            }
-            MedicineDataController.Instance.DeleteByID(medicine.MedicineData.Id);
         }
 
         private bool IsSelected(object obj)
